Exercicio 1 corrigido

+++ aula_01/src/aula_01/Ex01.java
@@ -8,15 +8,15 @@ public class Ex01 {
 	public static void main(String[] args) {
 	DecimalFormat df = new DecimalFormat("#,##,##0.00"); 
 	Scanner leia = new Scanner (System.in);   
-	double salario, abono;
+	float salario, abono;
 	String nome;
 	
 	System.out.println("Digite seu nome: ");
 	nome = leia.nextLine();
 	System.out.println("Digite seu Salario: ");
-	salario= leia.nextDouble();
+	salario= leia.nextFloat();
 	System.out.println("Digite seu Abono:  ");
-	abono= leia .nextDouble();
+	abono= leia .nextFloat();
 	System.out.println("Novo Salario: " + df.format(salario + abono));
 	}
